+++ tests/test_test_azure_parquet_remote.py
@@ -8,8 +8,6 @@ from datacontract.data_contract import DataContract
 
 logging.basicConfig(level=logging.INFO, force=True)
 
-datacontract = "fixtures/azure-parquet-remote/datacontract.yaml"
-
 load_dotenv(override=True)
 
 
@@ -19,10 +17,25 @@ load_dotenv(override=True)
     or os.environ.get("DATACONTRACT_AZURE_CLIENT_SECRET") is None,
     reason="Requires DATACONTRACT_AZURE_TENANT_ID, DATACONTRACT_AZURE_CLIENT_ID, and DATACONTRACT_AZURE_CLIENT_SECRET to be set",
 )
-def _test_test_azure_parquet_remote():
-    data_contract = DataContract(data_contract_file=datacontract)
+def test_test_azure_parquet_remote():
+    data_contract = DataContract(data_contract_file="fixtures/azure-parquet-remote/datacontract.yaml")
 
     run = data_contract.test()
 
     print(run)
+    assert run.result == "passed"
+
+
+@pytest.mark.skipif(
+    os.environ.get("DATACONTRACT_AZURE_TENANT_ID") is None
+    or os.environ.get("DATACONTRACT_AZURE_CLIENT_ID") is None
+    or os.environ.get("DATACONTRACT_AZURE_CLIENT_SECRET") is None,
+    reason="Requires DATACONTRACT_AZURE_TENANT_ID, DATACONTRACT_AZURE_CLIENT_ID, and DATACONTRACT_AZURE_CLIENT_SECRET to be set",
+)
+def test_test_azure_json_remote():
+    data_contract = DataContract(data_contract_file="fixtures/azure-json-remote/datacontract.yaml")
+
+    run = data_contract.test()
+
+    print(run.pretty())
     assert run.result == "passed"