+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Introduced optional dependencies/extras for significantly faster installation times. (#213)
 - Added delta-lake as an additional optional dependency
+- support `GOOGLE_APPLICATION_CREDENTIALS` as variable for connecting to bigquery in `datacontract test`
+- better support bigqueries `type` attribute, don't assume all imported models are tables
 
 ### Fixed
 - `datacontract export --format avro` fixed array structure (#243)